refactored GoogleDocIcon into component
