+++ src/components/common/GoogleDocIcon/index.js
@@ -1,0 +1,3 @@
+import GoogleDocIcon from './GoogleDocIcon';
+
+export default GoogleDocIcon;
