+++ src/components/common/GoogleDocIcon/style.css
@@ -1,0 +1,9 @@
+.google-doc-icon {
+  height: 20px;
+  margin-left: 10px;
+  cursor: pointer;
+}
+
+.google-doc-icon.inactive {
+  filter: grayscale(100%);
+}
